Integrate recommended book to review page
